[#163615731] refactor import statements
